rename server file
