+++ tests/adminserver.test.ts
@@ -2,8 +2,8 @@ import { DBOS, StatusString } from '../src';
 import { DBOSConfig } from '../src/dbos-executor';
 import { WorkflowQueue } from '../src';
 import { generateDBOSTestConfig, queueEntriesAreCleanedUp, setUpDBOSTestDb } from './helpers';
-import { QueueMetadataResponse } from '../src/httpServer/server';
-import { HealthUrl, WorkflowQueuesMetadataUrl, WorkflowRecoveryUrl } from '../src/httpServer/server';
+import { QueueMetadataResponse } from '../src/adminserver';
+import { HealthUrl, WorkflowQueuesMetadataUrl, WorkflowRecoveryUrl } from '../src/adminserver';
 import { globalParams, sleepms } from '../src/utils';
 import { Client } from 'pg';
 import { step_info } from '../schemas/system_db_schema';
